Migrated some real-life testing of Subversion interface.
